backend/pyproject.toml
[project]
name = "polish-football-tracker-backend"
version = "0.1.0"
description = "FastAPI backend for Polish Footballers Abroad Tracker"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Polish Tracker Team" }]

dependencies = [
    # FastAPI core
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "supabase>=2.10.0",

    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",

    # HTTP Client for RapidAPI
    "httpx>=0.28.0",

    # Rate Limiting
    "slowapi>=0.1.9",

    # Logging
    "structlog>=24.4.0",

    # Payments
    "stripe>=11.0.0",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # CORS & Middleware
    "starlette>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
